Fix bug with URL encoded params in endpoints
The manual construction of the struct url.URL was considering the URL
parameters as a part of the URLs's path and this was causing encoding
issues on some characters.

As an example, if a pod had the `prometheus.io/path` annotation set to
`/metrics?format=prometheus` we would end up querying
`/metrics%3Fformat=prometheus`.