Change delta calcaulator defaults (#52)
Changes the values for the telemetry emitter delta calculator expiration
age and expiration check interval from 30 seconds to 5 minutes. Also
adds the following configuration options for setting each setting:

- telemetry_emitter_delta_expiration_age
- telemetry_emitter_delta_expiration_check_interval

This fixes the issue of missing datapoints for counters when the scrape
interval was taking longer than the delta expiration check interval and age.