Update how histogram&summary metrics are reported (#54)
As New Relic is improving its support for Prometheus native format, we
could change the way how we report Summary and Histogram metric types
and do it in less disruptive way, meaning that we send them as they come
from Prometheus, without adding custom suffixes or calculating
percentiles from the integration side.

These are the details of the current implementation:

- xxx_bucket is ingested as a Counter (converted to a delta rather than
accumulative). Metrics containing dimension "le" == '+Inf' are also
sent, we don't omit them.
- xxx_count is passed through as a Counter (converted to a delta rather than accumulative).
- xxx_total passed through as a Counter (converted to a delta rather than accumulative).
- xxx_sum is passed through as a NR Summary converted to a delta rather
than accumulative, Min and Max here should be NaN. Its value can be negative.
The count field of the summary should be 1.
- For Prometheus summary metrics, we report the quantile as a dimension
and we don't add a "percentile" dimension.